cluster issuer and secrets

+++ pkg/ingress.sh
@@ -1,4 +1,4 @@
-# 💡 Caddy Ingress
+#💡 Caddy Ingress
 function ensure_ingress_caddy {
     # ❗ This ingress can't be used for sticky sessions
     # https://github.com/caddyserver/ingress/issues/74#issuecomment-962909479
@@ -18,20 +18,10 @@ function ensure_ingress_caddy {
         --atomic caddy caddy-ingress-controller >"$m"
     sed -i '/targetPort: http$/a \ \ \ \ \ \ nodePort: 30080' "$m"
     sed -i '/targetPort: https$/a \ \ \ \ \ \ nodePort: 30443' "$m"
-    shw add_namespace "caddy-system" "$m"
+    import render_namespace
+    shw render_namespace "caddy-system" "$m"
     shw kubectl apply -f "$m"
     ok "Caddy deployed" "Logs: $0 log caddy"
-}
-
-function add_namespace {
-    local h && h="$(cat "$2")"
-    echo -e "---
-apiVersion: v1
-kind: Namespace
-metadata:
-  name: $1
-$h
-" >"$2"
 }
 
 T_CADDY_VALS="
@@ -44,20 +34,24 @@ ingressController:
     proxyProtocol: true
 "
 
-# 💡 Ninx Ingress: A real ingresss which handles layer 7 based routing
-function ensure_ingress_nginx_with_certmgr {
+#💡 Ninx Ingress: A real ingresss which handles layer 7 based routing
+function ensure_ingress_nginx {
     # supports sticky sessions
-    # rm: Delete ingress and cert-manager
-    local l d="./deploys/nginx" && mkdir -p "$d"
-    local m="$d/manifest.yaml"
-    test "${1:-}" = "rm" && {
+    local m l d="./deploys/nginx" rm=false && m="$d/manifest.yaml"
+
+    while [[ -n "${1:-}" ]]; do case "$1" in
+        --rm | rm) rm=true ;; # Remove ingress
+        *) die "Unsupported" "$1" ;;
+        esac && shift; done
+
+    if $rm; then
         l="$(shw helm list -A)"
         grep -q ingress-nginx <<<"$l" && shw helm delete ingress-nginx -n ingress-nginx
-        grep -q cert-manager <<<"$l" && shw helm delete cert-manager -n cert-manager
-        ok "No ingress-nginx and cert-manager in your cluster"
+        ok "No more ingress-nginx in your cluster"
         return
-    }
-    test -z "$EMAIL" && die "EMAIL not set" "Set \$EMAIL, required for SSL"
+    fi
+
+    mkdir -p "$d"
 
     ok "Adding nginx"
     echo -e "$T_NGINX_VALS" >"$d/values.yaml"
@@ -65,16 +59,12 @@ function ensure_ingress_nginx_with_certmgr {
     shw helm upgrade --install ingress-nginx ingress-nginx \
         --repo https://kubernetes.github.io/ingress-nginx \
         --namespace ingress-nginx --create-namespace -f "$d/values.yaml"
+    ok "Nginx Ingress installed\n${L}Ensure also cert_manager and you can run 'test_http_svc_nginx'$O"
 
-    ok "Adding cert manager"
-    shw helm repo add jetstack https://charts.jetstack.io
-    shw helm upgrade --install --namespace cert-manager --create-namespace --set installCRDs=true cert-manager jetstack/cert-manager
-    r_certmgr "$EMAIL" >"$d/certmgr.yaml"
-    shw kubectl apply -f "$d/certmgr.yaml"
-    ok "Nginx Ingress and CertMgr using Lets encrypt installed\n${L}You can now run 'test_http_svc_nginx'$O"
 }
 
 T_NGINX_VALS=$(
+    # https://kubernetes.github.io/ingress-nginx/user-guide/nginx-configuration/configmap/
     cat <<'EOF'
 controller:
   kind: DaemonSet
@@ -84,44 +74,24 @@ controller:
       http: 30080
       https: 30443
   config:
-    # https://kubernetes.github.io/ingress-nginx/user-guide/nginx-configuration/configmap/
     use-proxy-protocol: "true"
     use-forwarded-headers: "true" # when others set XForwFor we take it
 
 EOF
 )
-
-r_certmgr() {
-    local email="${1:?Require email}"
-    cat <<EOF
-apiVersion: cert-manager.io/v1
-kind: ClusterIssuer
-metadata:
-  name: letsencrypt-prod
-  namespace: cert-manager
-spec:
-  acme:
-    email: $email
-    server: https://acme-v02.api.letsencrypt.org/directory
-    privateKeySecretRef:
-      name: letsencrypt-prod-account-key
-    solvers:
-    - http01:
-        ingress:
-          class: nginx
-EOF
-}
-
+#💡 For an actual service to be reachable via the ingress
 function render_ingress_nginx {
+    # ssl only  via letsencrypt-prod currently
     retval_=''
-    local name="" hostname="" stk=false ssl=true
-    while [[ -n "${1:-}" ]]; do
-        case "$1" in
-        h=* | hostname=*) hostname="${1##*=}" && shift ;;
-        n=* | name=*) name="${1##*=}" && shift ;;
-        -n | --nossl) ssl=false && shift ;;
-        -s | --sticky-sessions) stk=true && shift ;;
-        *) shift ;;
+    local name="" namespace=default hostname="" stk=false ssl=true
+    for arg in "$@"; do
+        case "$arg" in
+        N=* | namespace=*) namespace="${arg#*=}" ;;
+        h=* | hostname=*) hostname="${arg#*=}" ;;
+        n=* | name=*) name="${arg#*=}" ;;
+        -n | --nossl) ssl=false ;;
+        -s | --sticky-sessions) stk=true ;;
+        *) out "not processed: $arg" ;;
         esac
     done
     test -z "$name" && die "name not set" "💡 Use n <name>"
@@ -130,11 +100,9 @@ function render_ingress_nginx {
     if [[ $ssl == true ]] || [[ $stk == true ]]; then
         a='annotations:\n'
         $ssl && a=''$a'    cert-manager.io/cluster-issuer: "letsencrypt-prod"\n'
-        $ssl && a=''$a'    kubernetes.io/tls-acme: "true"\n'
         $stk && a=''$a'    kubernetes.io/tls-acme: "true"\n'
         $stk && a=''$a'    nginx.ingress.kubernetes.io/affinity: "cookie"\n'
         $stk && a=''$a'    nginx.ingress.kubernetes.io/session-cookie-name: "route"\n'
-        $stk && a=''$a'    nginx.ingress.kubernetes.io/session-cookie-expires: "172800"\n'
         $stk && a=''$a'    nginx.ingress.kubernetes.io/session-cookie-max-age: "172800"\n'
     fi
     retval_=$(
@@ -144,6 +112,7 @@ apiVersion: networking.k8s.io/v1
 kind: Ingress
 metadata:
   name: $name
+  namespace: $namespace
   $a
 spec:
   ingressClassName: nginx
@@ -166,3 +135,5 @@ EOF
     )
     echo -e "$retval_"
 }
+
+false && . ./tools.sh && . ./svc.sh || true
